Migrated all files reading/write to use System.property("user.dir")
This was done to handle issues when writing to resources
files. It is not really possible to write to files wrapped
in jar and therefore we need files located in the home directory
of the exe. This allows us to have local files that can be
utilized to read and write with.

Added logic for creating folders if they don't already exist.